Test cases now generated in engine.
Test case generation is done the exact same way for each problem module so I moved it to the engine.

For this to work and for the engine to import abitrary problem modules, I defined a constant TEST_CASE_TYPE_ENUM that will be available in every problem module so the engine can iterate over the test case types during generation.